blub

+++ product-service/product-service/src/main/java/de/hska/iwi/vslab/productservice/dao/ProductRepo.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import de.hska.iwi.vslab.productservice.dataobject.Product;
 
-public interface ProductRepo extends CrudRepository<Product, Long> {
+public interface ProductRepo extends CrudRepository<Product, Integer> {
 
 }
